requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
motor==3.7.1
python-multipart==0.0.20
python-dotenv==1.1.1
opencv-python==4.12.0.88
deepface==0.0.93
tf_keras
numpy>=2.0,<2.3.0
pillow==11.3.0
pymongo==4.13.2
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
